requirements.txt
django
django-rest-framework
drf_yasg
setuptools
pre-commit
